Enable stream scanning
This stream scanning implementation is
added "inside" the existing string
based Parse functions to ensure
backwards compatibility and to reuse
the existing test suite.

The reason for this change is to
reduce the memory usage of the parser
for stream based sources, mostly files.
Rather than having to read an entire
file into memory it can instead be
streamed in chunks.